apps/domain-manager/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
dnspython==2.7.0
